+++ TiendaBackEnd/Logica/logica-DTOFactura.cs
@@ -18,7 +18,7 @@ namespace Logica
             return op.seleccionarFacturaDTO();
         }
 
-        public DTO_Factura MostrarFacturaPorId(decimal codigo)
+        public DTO_Factura MostrarFacturaPorId(int codigo)
         {
             return MostrarFacturas().Where(pro => pro.FAC_COD == codigo).SingleOrDefault();
         }
